refactor(genesis landing): adding custom hook
