+++ src/hooks/useScrollPosition.js
@@ -1,0 +1,20 @@
+import { useState, useEffect } from 'react';
+
+export default function useScrollPosition() {
+  const [scrollPosition, setScrollPosition] = useState(0)
+
+  const handleScroll = () => {
+    const position = window.pageYOffset
+    setScrollPosition(position)
+  }
+
+  useEffect(() => {
+    window.addEventListener('scroll', handleScroll)
+
+    return () => {
+      window.removeEventListener('scroll', handleScroll)
+    }
+  }, [])
+
+  return scrollPosition;
+}
